LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/import-scripts-updated-flag.https-expected.txt

CONSOLE MESSAGE: line 30: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'registration.unregister')

FAIL initialize global state assert_unreached: unregister and register should not fail: Error: assert_equals: expected (string) "root" but got (object) null Reached unreachable code
FAIL import script previously imported at worker evaluation time promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'registration.installing')"
FAIL import script previously imported at worker install time promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'registration.installing')"
FAIL import script not previously imported promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'registration.installing')"
